backend/requirements.txt
fastapi
uvicorn[standard]
ultralytics
opencv-python
pillow
python-multipart
google-genai